Simplified Trait code, specifically the Level type
Since Level is really just an Int, decided to change this to a type synonym and worry about making a distinction at display time. This means a lot of typeclass instance code was removed from this module. And also the display functions needed to be changed. Also, added a new Trait type to combine a trait description string and a Level.